Adjusted the deserializers and added init binders to process UNumber related classes.
